+++ packages/svelte/scripts/copy-svelte.mjs
@@ -3,5 +3,8 @@ import { copyFileSync } from 'node:fs';
 import { URL } from 'node:url';
 
 for (const file of ['Trans.svelte', 'TransNodes.svelte', 'Trans.svelte.d.ts']) {
-  copyFileSync(new URL(`../src/${file}`, import.meta.url), new URL(`../dist/${file}`, import.meta.url));
+  copyFileSync(
+    new URL(`../src/${file}`, import.meta.url),
+    new URL(`../dist/${file}`, import.meta.url),
+  );
 }