+++ packages/svelte/src/Trans.svelte.d.ts
@@ -1,11 +1,12 @@
 import { SvelteComponent } from 'svelte';
-import type { Params, Verbaly } from 'verbaly';
+import type { Params, RichLink, Verbaly } from 'verbaly';
 
 export interface TransProps {
   id: string;
   values?: Params;
   instance?: Verbaly;
   richTags?: string[];
+  links?: Record<string, RichLink>;
 }
 
 export default class Trans extends SvelteComponent<TransProps> {}